+++ src/Redmine/Api/IssueStatus.php
@@ -34,7 +34,7 @@ class IssueStatus extends AbstractApi
 
     /**
      * @deprecated v2.9.0 Use fromHttpClient() instead.
-     * @see Group::fromHttpClient()
+     * @see IssueStatus::fromHttpClient()
      *
      * @param Client|HttpClient $client
      */